final version for geekathon
not working totally as expected because of AI lack of training on tracing points. the trips generated are not always good, as sometimes the destination is not completely reached and in other cases the trip segmentation is not a direct line as the trace sends you back sometimes. Apart from that, it is working properly, selecting destination on map, user preferences and so on. 2 to 5 trips are generated for the user and can be visualized on the interactive map